saleappv2/requirements.txt
Babel==2.11.0
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.30.0
colorama==0.4.6
crypto==1.4.1
cryptography==39.0.2
Flask==2.2.2
Flask-Admin==1.6.0
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
greenlet==1.1.3.post0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
Naked==0.1.32
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.6
PyYAML==6.0
requests==2.28.2
rsa==4.9
shellescape==3.8.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.42
twilio==7.16.4
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.10.0
